software: Add drivers, headers, and software tests
Consolidates C drivers and public headers for all peripherals and
system IPs added in this series. Adds a software test suite with
per-IP test cases and a CI workflow.

+++ software/driver/esm.c
@@ -6,7 +6,7 @@
 
 #include "esm.h"
 
-int esm_init(struct esm_driver *driver, unsigned int base_address)
+int esm_init(struct esm_driver *driver, unsigned long base_address)
 {
 	driver->regs = (volatile struct esm_regs *)base_address;
 	driver->bank_count = (driver->regs->info >> 9) & 0x7Fu;
