+++ software/include/plic.h
@@ -15,7 +15,7 @@ struct plic_driver {
 };
 
 
-int plic_init(struct plic_driver *driver, unsigned int base_address);
+int plic_init(struct plic_driver *driver, unsigned long base_address);
 int plic_irq_enable(struct plic_driver *driver, unsigned int number);
 int plic_irq_disable(struct plic_driver *driver, unsigned int number);
 int plic_irq_claim(struct plic_driver *driver, unsigned int number);
